+++ src/main/java/br/edu/ifsp/associacaogaia/repository/ArtesaoRepository.java
@@ -3,5 +3,5 @@ package br.edu.ifsp.associacaogaia.repository;
 import br.edu.ifsp.associacaogaia.model.Artesao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArtesaoRepository extends JpaRepository<Artesao, Integer> {
+public interface ArtesaoRepository extends JpaRepository<Artesao, Long> {
 }
